각 도메인 pk 자료형 변환 이슈로 인한 컴파일 오류 해결

+++ src/main/java/com/team/berp/client/service/ClientService.java
@@ -8,20 +8,20 @@ import org.springframework.data.domain.Pageable;
 
 public interface ClientService {
     // 등록
-    Integer register(ClientViewDto dto, Employee employee);
+    Long register(ClientViewDto dto, Employee employee);
 
     // 수정
-    void update(Integer companyId, ClientViewDto dto, Employee employee);
+    void update(Long companyId, ClientViewDto dto, Employee employee);
 
     // 상세조회
-    ClientViewDto getById(Integer companyId);
+    ClientViewDto getById(Long companyId);
 
     // 리스트조회(검색, 페이징)
     Page<ClientViewDto> getList(CompanyType type, String keyword, String searchType, Pageable pageable);
 
     // 삭제
-    void delete(Integer companyId);
+    void delete(Long companyId);
 
     // 상태 변경
-    void changeUseYn(Integer companyId, String useYn);
+    void changeUseYn(Long companyId, String useYn);
 }
